에디터 - xml 파일 저장 기능 추가
드디어

+++ Assets/Scripts/Edit/Copy.cs
@@ -56,9 +56,9 @@ public class Copy : MonoBehaviour {
 	public bool visible = false;
 
 	[NonSerialized]
-	public float playerX = 0;
+	public int playerX = 0;
 	[NonSerialized]
-	public float playerY = 0;
+	public int playerY = 0;
 
 	[NonSerialized]
 	public string entityTag = "Enemy";
